Sequential page type generator
Easier to script, batch delete, etc.
You can run it multiple times and it keeps incrementing the count automatically.
I'll keep the faker dependency around since we'll likely use it for other purposes again.
Also fixed a bug in the system where it was doing the basename() checks in the wrong place.